Generate Dhall types for TwitterCard
Dhall types will help content authors avoid the usual "Why didn't this
render as expected?" troubles that is expected out of static website
generators.

This is an attempt at generating Dhall types with the following
objectives:

1) Site, Node and Page metadata may have different types based on the
   desired feature set. Having the types generated (during site creation
   or feature set change) helps with the ergonomics immensely.
2) Avoids maintaining complex types in Haskell and Dhall
   codebases.
3) Makes configuration extensible with layers. For example, a S3 deploy
   layer may add S3 specific options to the meta data configuration or a
   google analytics layer (ugh!) may add configuration options for the
   analytics profile id etc.,

I'm still wading through the Dhall code and examples, so all this may
still be throwaway code. For example, this effort has significant
parallels to the `dhall rewrite-with-schemas` command and perhaps it is
more judicious use that machinery instead of the custom `LetBound`
typeclass.

Regardless, I'm still stoked that I wrote some working Haskell code
solving a specific problem of relatively decent complexity after a
long time :)