Add `Module` to automatically infer nested types
The previous commits added a way to generate a Dhall file. However,
there was a ton of boilerplate with a hacky `LetBound` typeclass. Not
only was this unergonomic, it was also not type safe. The types provided
no evidence that the type declarations were in the context of a let
binding.

This commit adds a state monad `Module` that keeps track of bindings
in a `Dhall` file. The `evalModule` function extracts the bindings as
`Text`.

This uses the `Dhall.Diff` machinery to do the comparison with
rudimentary depth based gate keeper for fast negatives.

This commit leaves the code for `TwitterCard` as is for the succeeding
commits to clean up.

Next steps:

- [ ] Handle imports
- [ ] Handle defaults