Ensure dhall files can be generated with imports
Most of the code for Urai and extensions shall be written in Haskell.
Most of the code for Urai generated sites, however, I'd like them to be
written in Dhall. This requires the following:

* Good ergonomics
* Batteries-included library to make Urai related functions and types
  readily available
* Ability to add plugins to change the Dhall context
* Granular type system for user defined data and functions

In order to accomplish these goals, the strategy I have chosen is to
generate Dhall files through Haskell types. This commit allows
generation of Dhall types that include other Dhall types (also
generated).

This is still not robust enough to handle all situations. Nested imports
for example. However, this seems to solve the main use cases for Urai.

Next Steps:

- [ ] Generate defaults
- [ ] Feature set based generation